Update some technical details.
Fix a few descriptions that were inaccurate or outdated.

Convert `LIMIT 1` to `assert_single()` throughout the text and examples.

Remove some unnecessary commands from "code" and remove obsolete
commands from REPL examples.

Change the language to "computed" as opposed to "computable".

Update the enum syntax to use "." instead of a cast.

Update the output formatting to match beta3 (which should be more
representative of the stable version).